fix(build): erros de compilação Kotlin do run 34766810404
- GameDialogs (2x): 'resources' não resolve dentro de window?.apply — o
  receiver é android.view.Window, que NÃO expõe getResources(). Métricas
  agora lidas do context do Dialog antes do apply.
- VirtualGamepadView: Math.toDegrees(Double) recebia Float do atan2 do
  kotlin.math (overload Float→Float) → .toDouble() explícito.

+++ app/src/main/java/com/deivid22srk/restuff/game/VirtualGamepadView.kt
@@ -376,7 +376,7 @@ class VirtualGamepadView(
     private fun dpadMaskAt(pad: Pad, x: Float, y: Float): Int {
         val cx = pad.cx * viewW
         val cy = pad.cy * viewH
-        val angle = Math.toDegrees(atan2(y - cy, x - cx))
+        val angle = Math.toDegrees(atan2(y - cy, x - cx).toDouble())
         return when {
             angle >= 157.5 || angle < -157.5 -> BIT_DLEFT
             angle >= 112.5 -> BIT_DLEFT or BIT_DDOWN
